Fixed an overflow with explicit weights that was an old bug outside the scope of conformance bitstreams

+++ edge264_internal.h
@@ -188,7 +188,7 @@ typedef struct
 	union { int8_t refIdx4x4_eq[32]; v16qi refIdx4x4_eq_v[2]; }; // FIXME store on stack
 	int16_t DistScaleFactor[32]; // [refIdxL0]
 	union { int8_t implicit_weights[32][32]; v16qi implicit_weights_v[32][2]; }; // w1 for [ref0][ref1]
-	int8_t explicit_weights[3][64]; // [iYCbCr][LX][RefIdx]
+	int16_t explicit_weights[3][64]; // [iYCbCr][LX][RefIdx]
 	int8_t explicit_offsets[3][64];
 	union { uint8_t edge_buf[2016]; int64_t edge_buf_l[252]; v16qu edge_buf_v[126]; };
